requirements.txt
discord.py
aiohttp
parsedatetime
python-dateutil
jishaku
humanize
psutil
beautifulsoup4
lxml
wavelink
pillow
colorama
matplotlib
tabulate
asyncpg
lru_dict
pynacl
viper-lang
pytz